Linux (Ubuntu x64) port: Win32-on-POSIX shim, SDL2/GPU bring-up, in-process harness
Bring Enemy Nations building and running on Linux from the linux-build branch,
keeping the Windows/MSVC build intact (all changes gated #ifndef _WIN32 /
if(NOT MSVC)).

Toolchain/CMake:
- Branch MSVC-only blocks in root, enations_latest/src, and wind22 CMakeLists
  with if(NOT MSVC): gcc flags, find_package(SDL2)+pkg-config, drop .lib names,
  DLL-copy, bundled includes, lastplnt.rc, PCH; gate _ITERATOR_DEBUG_LEVEL.

Win32 shim (no call-site rewrites):
- win32_compat.h/.cpp: fixed-width LP64 typedefs, calling-convention macros,
  sync/threads, Interlocked* (atomics), time, files/modules, registry over an
  ini, real GetClientRect/MoveWindow/SetWindowPos, cursor/screen-metric globals.
- winshim/ case-variant Windows headers for case-sensitive includes.
- win32_legacy_stubs.cpp: CAppPalette::GetColorValue, palette/CColorBuffer.
- linux_main.cpp (main->WinMain), win32_resources_linux.cpp (LoadStringA table),
  vdmplay_stubs.cpp.

LP64 correctness:
- fixpoint.h Fixed union long->int (32-bit on all platforms) + (long) casts on
  arithmetic operators so they bind the raw Fixed(long) ctor.
- mmio ReadLong reads int32_t and sign-extends; sprite.h/.cpp long->LONG.

SDL2/GPU runtime:
- Default to the SDL2/GPU renderer on Linux (RenderBackend.h); fullscreen.
- SDL2Terrain tile path uses '/' on Linux so stb_image fopen succeeds (530 tiles).
- Main-menu button text fit; create-status sprintf_s fixes.

Harness:
- harness/control_socket.cpp: in-process TCP control socket (EN_HARNESS=1) with
  shot/shotid/click/clickid/dblclickid/move/moveid/key/text/wins/raise/quit.

Build output dir build-linux/ added to .gitignore.

+++ enations_latest/src/DlgReg.cpp
@@ -23,7 +23,7 @@ std::string GetDefaultApp ( char const *pExt, char const *pDef, char const *pCmd
 		return sRtn;
 
 	// read it in
-	unsigned long iLen = 256;
+	DWORD iLen = 256;
 	DWORD dwTyp;
 	if (RegQueryValueEx (key, "", NULL, &dwTyp, (BYTE*) cmd, &iLen) != ERROR_SUCCESS)
 		return sRtn;
